refactor(contentscript): exports changed
