+++ .github/workflows/publish.yml
@@ -24,4 +24,4 @@ jobs:
       - name: Compile Extension
         run: |
           npm install
-          npm run start
+          npm run build
